requirements.txt
eel>=0.16.0
google-genai>=1.0.0
python-dotenv>=1.0.0
numpy>=1.24.0
